requirements.txt
# Core scientific computing
numpy>=1.24.0,<2.0.0
scipy>=1.10.0,<2.0.0

# Plotting
matplotlib>=3.7.0,<4.0.0

# Configuration
pyyaml>=6.0,<7.0.0

# Data handling (optional but included for future extensions)
pandas>=2.0.0,<3.0.0

